Handle sampling stacks with methods implemented using native code
A frame representing a method implemented using native code needs to
be handled in a special way:

* we need to use a different API to get the method name
* we need to use the path and line number of its direct/indirect Ruby
  caller (if we want to match the Ruby backtrace APIs at least)

To implement this, I further extended our custom `rb_profile_frames`
to support flagging a method as being a Ruby frame or not.

In terms of memory we could probably pack things a bit more (e.g.
not use a space-wasting array of bool) but I'll leave such
optimizations for later.

Finally, the `rb_profile_frame_method_name` method only returns
the actual method name for a native method on Ruby 3.0+.

To support older Rubies, a further commit will introduce a few more
copy-pasted backported code.