Add alternative `rb_profile_frames` implementation for Ruby 2.2 and below
The `rb_profile_frames` function changed quite a bit between Ruby 2.2
and 2.3.

Since the change was quite complex I opted not to try to extend support
to Ruby 2.2 and below using the same custom function, and instead I
started anew from the Ruby 2.2 version of the function, applying some
of the same fixes that we have for the modern version.

This gets us closer to a green profiler test suite on 2.2 and 2.1
(there's at least one other compatibility fix needed elsewhere which
will come in a later commit).