Refactor `build_request` and `_native_do_export` to avoid leaking libddprof objects
The previous version of the code went:

1. Create libddprof exporter
2. Convert Ruby VALUE objects into libddprof equivalents
3. Fire off the request

In the happy path, this worked fine. But if in step 2 any exception
was raised (because some object was `nil` or of an unexpected type),
then the exporter would be leaked.

To fix this, I've flipped the code around so that the order now goes:

1. Convert Ruby VALUE objects into libddprof equivalents
2. Create libddprof exporter
3. Fire off the request

Thus, we now do all of the work that may raise exceptions up front,
and only then do we start calling libddprof APIs which allocate dynamic
memory.

(I did slightly move the code around to minimize the diff, and that's
why `build_request` actually became `_native_do_export`, and the old
code in `_native_do_export` became the new `perform_export`).