Import `interesting_backtrace_helper.rb` from backtracie gem and use it for testing the stack collector
(As the creator and sole copyright owner of that file in the
`backtracie` gem, taken from
<https://github.com/ivoanjo/backtracie/blob/1f6116ff5b07e87fe3e5be95a22c6a8e1210de8b/spec/unit/interesting_backtrace_helper.rb>
I am relicensing the file and its copyrights to belong to Datadog
similarly to the rest of the dd-trace-rb codebase).

This file is used as a single mega-complex test for the stack
sampling code, and is quite useful to spot regressions.